requirements.txt
tqdm==4.28.1
matplotlib==3.1.1
Keras==2.2.4
pandas==0.23.4
scipy==1.3.1
pygame==1.9.6
opencv_python==4.0.0.21
setuptools==41.4.0
Pillow==6.2.0
numpy==1.17.2
scikit_learn==0.21.3
tensorflow==1.15rc2
imageio==2.6.1
imagehash==4.0

